refactor(NotifyMessage): 코드 및 소제목 추가
- Converter 사용을 위해 코드로 찾을 수 있는 메서드 추가로 정의했습니다.
